apache_avro

Function write_avro_datum_ref

pub fn write_avro_datum_ref<T: Serialize, W: Write>(
    schema: &Schema,
    data: &T,
    writer: &mut W,
) -> AvroResult<usize>

Write the referenced Serializeable object to the provided Write object. Returns a result with the number of bytes written.

NOTE: This function has a quite small niche of usage and does NOT generate headers and sync markers; use append_ser to be fully Avro-compatible if you don’t know what you are doing, instead.