apache_avro::headers

Trait HeaderBuilder

pub trait HeaderBuilder {
    // Required method
    fn build_header(&self) -> Vec<u8> ;
}

This trait represents that an object is able to construct an Avro message header.

It is implemented for some known header types already. If you need a header type that is not already included here, then you can create your own struct and implement this trait.

Required Methods

fn build_header(&self) -> Vec<u8>