apache_avro::rabin

Struct Rabin

pub struct Rabin { /* private fields */ }

Implementation of the Rabin fingerprint algorithm using the Digest trait as described in [schema fingerprints].

The digest is returned as the 8-byte little-endian encoding of the Rabin hash. This is what is used for Avro [single object encoding]

use apache_avro::rabin::Rabin;
use digest::Digest;
use hex_literal::hex;

// create the Rabin hasher
let mut hasher = Rabin::new();

// add the data
hasher.update(b"hello world");

// read hash digest and consume hasher
let result = hasher.finalize();

assert_eq!(result[..], hex!("60335ba6d0415528"));

To convert the digest to the commonly used 64-bit integer value, you can use the i64::from_le_bytes() function

let i = i64::from_le_bytes(result.try_into().unwrap());

assert_eq!(i, 2906301498937520992)

single object encoding schema fingerprints