apache_avro::reader::datum

Struct GenericDatumReaderBuilder

pub struct GenericDatumReaderBuilder<'s, S: State = Empty> { /* private fields */ }

Use builder syntax to set the inputs and finish with build().

Implementations

impl<'s, S: State> GenericDatumReaderBuilder<'s, S>

pub fn build(self) -> AvroResult<GenericDatumReader<'s>>
where S: IsComplete,

Finishes building and performs the requested action.

pub fn resolved_writer_schemata( self, value: ResolvedSchema<'s>, ) -> GenericDatumReaderBuilder<'s, SetResolvedWriterSchemata<S>>
where S::ResolvedWriterSchemata: IsUnset,

Optional (Some / Option setters). Already resolved schemata that will be used to resolve references in the writer’s schema.

pub fn maybe_resolved_writer_schemata( self, value: Option<ResolvedSchema<'s>>, ) -> GenericDatumReaderBuilder<'s, SetResolvedWriterSchemata<S>>
where S::ResolvedWriterSchemata: IsUnset,

Optional (Some / Option setters). Already resolved schemata that will be used to resolve references in the writer’s schema.

pub fn reader_schema( self, value: &'s Schema, ) -> GenericDatumReaderBuilder<'s, SetReaderSchema<S>>
where S::ReaderSchema: IsUnset,

Optional (Some / Option setters). The schema that will be used to resolve the value to conform the the new schema.

pub fn maybe_reader_schema( self, value: Option<&'s Schema>, ) -> GenericDatumReaderBuilder<'s, SetReaderSchema<S>>
where S::ReaderSchema: IsUnset,

Optional (Some / Option setters). The schema that will be used to resolve the value to conform the the new schema.

pub fn resolved_reader_schemata( self, value: ResolvedSchema<'s>, ) -> GenericDatumReaderBuilder<'s, SetResolvedReaderSchemata<S>>
where S::ResolvedReaderSchemata: IsUnset,

Optional (Some / Option setters). Already resolved schemata that will be used to resolve references in the reader’s schema.

pub fn maybe_resolved_reader_schemata( self, value: Option<ResolvedSchema<'s>>, ) -> GenericDatumReaderBuilder<'s, SetResolvedReaderSchemata<S>>
where S::ResolvedReaderSchemata: IsUnset,

Optional (Some / Option setters). Already resolved schemata that will be used to resolve references in the reader’s schema.

impl<'s, S: State> GenericDatumReaderBuilder<'s, S>

pub fn writer_schemata( self, schemata: Vec<&'s Schema>, ) -> AvroResult<GenericDatumReaderBuilder<'s, SetResolvedWriterSchemata<S>>>
where S::ResolvedWriterSchemata: IsUnset,

Set the schemata that will be used to resolve any references in the writer’s schema.

This is equivalent to .resolved_writer_schemata(ResolvedSchema::new_with_schemata(schemata)?). If you already have a ResolvedSchema, use that function instead.

pub fn reader_schemata( self, schemata: Vec<&'s Schema>, ) -> AvroResult<GenericDatumReaderBuilder<'s, SetResolvedReaderSchemata<S>>>
where S::ResolvedReaderSchemata: IsUnset, S::ReaderSchema: IsSet,

Set the schemata that will be used to resolve any references in the reader’s schema.

This is equivalent to .resolved_reader_schemata(ResolvedSchema::new_with_schemata(schemata)?). If you already have a ResolvedSchema, use that function instead.

This function can only be called after the reader schema is set.