apache_avro::schema

Enum SchemaKind

pub enum SchemaKind {

    Null,
    Boolean,
    Int,
    Long,
    Float,
    Double,
    Bytes,
    String,
    Array,
    Map,
    Union,
    Record,
    Enum,
    Fixed,
    Decimal,
    BigDecimal,
    Uuid,
    Date,
    TimeMillis,
    TimeMicros,
    TimestampMillis,
    TimestampMicros,
    TimestampNanos,
    LocalTimestampMillis,
    LocalTimestampMicros,
    LocalTimestampNanos,
    Duration,
    Ref,

}

Variants

Null

A null Avro schema.

Boolean

A boolean Avro schema.

Int

An int Avro schema.

Long

A long Avro schema.

Float

A float Avro schema.

Double

A double Avro schema.

Bytes

A bytes Avro schema.

Bytes represents a sequence of 8-bit unsigned bytes.

String

A string Avro schema.

String represents a unicode character sequence.

Array

An array Avro schema.

All items will have the same schema.

Map

A map Avro schema.

Keys are always a Schema::String and all values will have the same schema.

Union

A union Avro schema.

Record

A record Avro schema.

Enum

An enum Avro schema.

Fixed

A fixed Avro schema.

Decimal

Logical type which represents Decimal values.

The underlying type is serialized and deserialized as Schema::Bytes or Schema::Fixed.

BigDecimal

Logical type which represents Decimal values without predefined scale.

The underlying type is serialized and deserialized as Schema::Bytes

Uuid

A universally unique identifier, annotating a string, bytes or fixed.

Date

Logical type which represents the number of days since the unix epoch.

Serialization format is Schema::Int.

TimeMillis

The time of day in number of milliseconds after midnight.

This type has no reference to any calendar, time zone or date in particular.

TimeMicros

The time of day in number of microseconds after midnight.

This type has no reference to any calendar, time zone or date in particular.

TimestampMillis

An instant in time represented as the number of milliseconds after the UNIX epoch.

TimestampMicros

An instant in time represented as the number of microseconds after the UNIX epoch.

TimestampNanos

An instant in time represented as the number of nanoseconds after the UNIX epoch.

LocalTimestampMillis

An instant in localtime represented as the number of milliseconds after the UNIX epoch.

LocalTimestampMicros

An instant in local time represented as the number of microseconds after the UNIX epoch.

LocalTimestampNanos

An instant in local time represented as the number of nanoseconds after the UNIX epoch.

Duration

An amount of time defined by a number of months, days and milliseconds.

Ref

A reference to another schema.

Implementations

impl SchemaKind

pub fn is_primitive(self) -> bool

pub fn is_named(self) -> bool

๐Deprecated since 0.22.0: Use Schema::is_named instead