apache_avro::schema

Struct RecordFieldBuilder

pub struct RecordFieldBuilder<S: State = Empty> { /* private fields */ }

Use builder syntax to set the inputs and finish with build().

Implementations

impl<S: State> RecordFieldBuilder<S>

pub fn build(self) -> RecordField
where S: IsComplete,

Finish building and return the requested object

pub fn name(self, value: impl Into<String>) -> RecordFieldBuilder<SetName<S>>
where S::Name: IsUnset,

Required.

Name of the field.

pub fn doc(self, value: Documentation) -> RecordFieldBuilder<SetDoc<S>>
where S::Doc: IsUnset,

Optional (Some / Option setters). Default: <Documentation as Default>::default().

Documentation of the field.

pub fn maybe_doc( self, value: Option<Documentation>, ) -> RecordFieldBuilder<SetDoc<S>>
where S::Doc: IsUnset,

Optional (Some / Option setters). Default: <Documentation as Default>::default().

Documentation of the field.

pub fn aliases(self, value: Vec<String>) -> RecordFieldBuilder<SetAliases<S>>
where S::Aliases: IsUnset,

Optional (Some / Option setters). Aliases of the field’s name. They have no namespace.

pub fn maybe_aliases( self, value: Option<Vec<String>>, ) -> RecordFieldBuilder<SetAliases<S>>
where S::Aliases: IsUnset,

Optional (Some / Option setters). Aliases of the field’s name. They have no namespace.

pub fn default(self, value: Value) -> RecordFieldBuilder<SetDefault<S>>
where S::Default: IsUnset,

Optional (Some / Option setters). Default value of the field. This value will be used when reading Avro datum if schema resolution is enabled.

pub fn maybe_default( self, value: Option<Value>, ) -> RecordFieldBuilder<SetDefault<S>>
where S::Default: IsUnset,

Optional (Some / Option setters). Default value of the field. This value will be used when reading Avro datum if schema resolution is enabled.

pub fn schema(self, value: Schema) -> RecordFieldBuilder<SetSchema<S>>
where S::Schema: IsUnset,

Required.

Schema of the field.

pub fn order(self, value: RecordFieldOrder) -> RecordFieldBuilder<SetOrder<S>>
where S::Order: IsUnset,

Optional (Some / Option setters). Default: RecordFieldOrder::Ignore.

Order of the field.

NOTE This currently has no effect.

pub fn maybe_order( self, value: Option<RecordFieldOrder>, ) -> RecordFieldBuilder<SetOrder<S>>
where S::Order: IsUnset,

Optional (Some / Option setters). Default: RecordFieldOrder::Ignore.

Order of the field.

NOTE This currently has no effect.

pub fn position(self, value: usize) -> RecordFieldBuilder<SetPosition<S>>
where S::Position: IsUnset,

Optional (Some / Option setters). Default: 0.

Position of the field in the list of field of its parent Schema

pub fn maybe_position( self, value: Option<usize>, ) -> RecordFieldBuilder<SetPosition<S>>
where S::Position: IsUnset,

Optional (Some / Option setters). Default: 0.

Position of the field in the list of field of its parent Schema

pub fn custom_attributes( self, value: BTreeMap<String, Value>, ) -> RecordFieldBuilder<SetCustomAttributes<S>>
where S::CustomAttributes: IsUnset,

Optional (Some / Option setters). Default: BTreeMap::new().

A collection of all unknown fields in the record field.

pub fn maybe_custom_attributes( self, value: Option<BTreeMap<String, Value>>, ) -> RecordFieldBuilder<SetCustomAttributes<S>>
where S::CustomAttributes: IsUnset,

Optional (Some / Option setters). Default: BTreeMap::new().

A collection of all unknown fields in the record field.