apache_avro::schema

Struct RecordSchemaBuilder

pub struct RecordSchemaBuilder<S: State = Empty> { /* private fields */ }

Use builder syntax to set the inputs and finish with build().

Implementations

impl<S: State> RecordSchemaBuilder<S>

pub fn build(self) -> RecordSchema
where S: IsComplete,

Finish building and return the requested object

pub fn name(self, value: Name) -> RecordSchemaBuilder<SetName<S>>
where S::Name: IsUnset,

Required.

The name of the schema

pub fn aliases(self, value: Aliases) -> RecordSchemaBuilder<SetAliases<S>>
where S::Aliases: IsUnset,

Optional (Some / Option setters). Default: <Aliases as Default>::default().

The aliases of the schema

pub fn maybe_aliases( self, value: Option<Aliases>, ) -> RecordSchemaBuilder<SetAliases<S>>
where S::Aliases: IsUnset,

Optional (Some / Option setters). Default: <Aliases as Default>::default().

The aliases of the schema

pub fn doc(self, value: Documentation) -> RecordSchemaBuilder<SetDoc<S>>
where S::Doc: IsUnset,

Optional (Some / Option setters). Default: <Documentation as Default>::default().

The documentation of the schema

pub fn maybe_doc( self, value: Option<Documentation>, ) -> RecordSchemaBuilder<SetDoc<S>>
where S::Doc: IsUnset,

Optional (Some / Option setters). Default: <Documentation as Default>::default().

The documentation of the schema

pub fn fields( self, value: Vec<RecordField>, ) -> RecordSchemaBuilder<SetFields<S>>
where S::Fields: IsUnset,

Required.

The set of fields of the schema

pub fn lookup( self, value: BTreeMap<String, usize>, ) -> RecordSchemaBuilder<SetLookup<S>>
where S::Lookup: IsUnset,

Required.

The lookup table maps field names to their position in the Vec of fields.

pub fn attributes( self, value: BTreeMap<String, Value>, ) -> RecordSchemaBuilder<SetAttributes<S>>
where S::Attributes: IsUnset,

Optional (Some / Option setters). Default: BTreeMap::new().

The custom attributes of the schema

pub fn maybe_attributes( self, value: Option<BTreeMap<String, Value>>, ) -> RecordSchemaBuilder<SetAttributes<S>>
where S::Attributes: IsUnset,

Optional (Some / Option setters). Default: BTreeMap::new().

The custom attributes of the schema