apache_avro::schema

Struct ResolvedSchema

pub struct ResolvedSchema<'s> { /* private fields */ }

Implementations

impl<'s> ResolvedSchema<'s>

pub fn get_schemata(&self) -> &[&'s Schema]

pub fn get_names(&self) -> &NamesRef<'s>

pub fn new(schema: &'s Schema) -> AvroResult<Self>

Resolve all references in this schema.

If some references are to other schemas, see ResolvedSchema::new_with_schemata.

pub fn new_with_schemata(schemata: Vec<&'s Schema>) -> AvroResult<Self>

Resolve all references in these schemas.

These schemas will be resolved in order, so references to schemas later in the list is not supported.

pub fn new_with_known_schemata<'n>( schemata_to_resolve: Vec<&'s Schema>, enclosing_namespace: &Namespace, known_schemata: &'n NamesRef<'n>, ) -> AvroResult<Self>

Creates ResolvedSchema with some already known schemas.

Those schemata would be used to resolve references if needed.