apache_avro

Module serde

Modules

bytes

Efficient (de)serialization of Avro bytes values.

bytes_opt

Efficient (de)serialization of optional Avro bytes values.

fixed

Efficient (de)serialization of Avro fixed values.

fixed_opt

Efficient (de)serialization of optional Avro fixed values.

slice

Efficient (de)serialization of Avro bytes/fixed borrowed values.

slice_opt

Efficient (de)serialization of optional Avro bytes/fixed borrowed values.

Traits

AvroSchema

Trait for types that serve as an Avro data model. Derive implementation available through derive feature. Do not implement directly! Implement AvroSchemaComponent to get this trait through a blanket implementation.

AvroSchemaComponent

Trait for types that serve as fully defined components inside an Avro data model. Derive implementation available through derive feature. This is what is implemented by the derive(AvroSchema) macro.

Functions

from_value

Interpret a Value as an instance of type D.

to_value

Interpret a serializeable instance as a Value.