apache_avro::serde

Module fixed

Efficient (de)serialization of Avro fixed values.

This module is intended to be used through the Serde with attribute. Use apache_avro::serde::fixed_opt for optional fixed values.

When used with different serialization formats, this is equivalent to serde_bytes.

See usage with below example:

#[derive(AvroSchema, Serialize, Deserialize)]
struct StructWithBytes {
    #[avro(with)]
    #[serde(with = "apache_avro::serde::bytes")]
    vec_field: Vec<u8>,

    #[avro(with = apache_avro::serde::fixed::get_schema_in_ctxt::<6>)]
    #[serde(with = "apache_avro::serde::fixed")]
    fixed_field: [u8; 6],
}

Functions

deserialize

get_record_fields_in_ctxt

Returns None

get_schema_in_ctxt

Returns Schema::Fixed(N) named serde_avro_fixed_{N}

serialize