apache_avro::serde

Trait AvroSchemaComponent

pub trait AvroSchemaComponent {
    // Required method
    fn get_schema_in_ctxt(
        named_schemas: &mut HashSet<Name>,
        enclosing_namespace: NamespaceRef<'_>,
    ) -> Schema;

    // Provided methods
    fn get_record_fields_in_ctxt(
        named_schemas: &mut HashSet<Name>,
        enclosing_namespace: NamespaceRef<'_>,
    ) -> Option<Vec<RecordField>> { ... }
    fn field_default() -> Option<Value> { ... }
}

Trait for types that serve as fully defined components inside an Avro data model.

This trait can be derived with #[derive(AvroSchema)] when the derive feature is enabled.

Implementation guide

Implementation for returning primitive types

When the schema you want to return is a primitive type (a type without a name), the function arguments can be ignored.

For example, you have a custom integer type:

// Make sure to implement `Serialize` and `Deserialize` to use the right serialization methods
pub struct U24([u8; 3]);
impl AvroSchemaComponent for U24 {
    fn get_schema_in_ctxt(_: &mut HashSet<Name>, _: NamespaceRef) -> Schema {
        Schema::Int
    }

    fn get_record_fields_in_ctxt(_: &mut HashSet<Name>, _: NamespaceRef) -> Option<Vec<RecordField>> {
        None // A Schema::Int is not a Schema::Record so there are no fields to return
    }

    fn field_default() -> Option<serde_json::Value> {
        // Zero as default value. Can also be None if you don't want to provide a default value
        Some(0u8.into())
    }
}

Passthrough implementation

To construct a schema for a type is “transparent”, such as for smart pointers, simply pass through the arguments to the inner type:

#[derive(Serialize, Deserialize)]
#[serde(transparent)] // This attribute is important for all passthrough implementations!
pub struct Transparent<T>(T);
impl<T: AvroSchemaComponent> AvroSchemaComponent for Transparent<T> {
    fn get_schema_in_ctxt(named_schemas: &mut HashSet<Name>, enclosing_namespace: NamespaceRef) -> Schema {
        T::get_schema_in_ctxt(named_schemas, enclosing_namespace)
    }

    fn get_record_fields_in_ctxt(named_schemas: &mut HashSet<Name>, enclosing_namespace: NamespaceRef) -> Option<Vec<RecordField>> {
        T::get_record_fields_in_ctxt(named_schemas, enclosing_namespace)
    }

    fn field_default() -> Option<serde_json::Value> {
        T::field_default()
    }
}

Implementation for complex types

When the schema you want to return is a complex type (a type with a name), special care has to be taken to avoid duplicate type definitions and getting the correct namespace.

Things to keep in mind:

If the fully qualified name already exists, return a Schema::Ref

Use the AvroSchemaComponent implementations to get the schemas for the subtypes

The ordering of fields in the schema must match with the ordering in Serde

Implement get_record_fields_in_ctxt as the default implementation has to be implemented with backtracking and a lot of cloning.

Even if your schema is not a record, still implement the function and just return None

Implement field_default() if you want to use #[serde(skip_serializing{,_if})].

pub struct Foo {
    one: String,
    two: i32,
    three: Option<Duration>
}

impl AvroSchemaComponent for Foo {
    fn get_schema_in_ctxt(named_schemas: &mut HashSet<Name>, enclosing_namespace: NamespaceRef) -> Schema {
        // Create the fully qualified name for your type given the enclosing namespace
        let name = Name::new_with_enclosing_namespace("Foo", enclosing_namespace).expect("Name is valid");
        if named_schemas.contains(&name) {
            Schema::Ref { name }
        } else {
            let enclosing_namespace = name.namespace();
            // Do this before you start creating the schema, as otherwise recursive types will cause infinite recursion.
            named_schemas.insert(name.clone());
            let schema = Schema::Record(RecordSchema::builder()
                .name(name.clone())
                .fields(Self::get_record_fields_in_ctxt(named_schemas, enclosing_namespace).expect("Impossible!"))
                .build()
            );
            schema
        }
    }

    fn get_record_fields_in_ctxt(named_schemas: &mut HashSet<Name>, enclosing_namespace: NamespaceRef) -> Option<Vec<RecordField>> {
        Some(vec![
            RecordField::builder()
                .name("one")
                .schema(String::get_schema_in_ctxt(named_schemas, enclosing_namespace))
                .build(),
            RecordField::builder()
                .name("two")
                .schema(i32::get_schema_in_ctxt(named_schemas, enclosing_namespace))
                .build(),
            RecordField::builder()
                .name("three")
                .schema(<Option<Duration>>::get_schema_in_ctxt(named_schemas, enclosing_namespace))
                .build(),
        ])
    }

    fn field_default() -> Option<serde_json::Value> {
        // This type does not provide a default value
        None
    }
}

Required Methods

fn get_schema_in_ctxt( named_schemas: &mut HashSet<Name>, enclosing_namespace: NamespaceRef<'_>, ) -> Schema

Get the schema for this component

Provided Methods

fn get_record_fields_in_ctxt( named_schemas: &mut HashSet<Name>, enclosing_namespace: NamespaceRef<'_>, ) -> Option<Vec<RecordField>>

Get the fields of this schema if it is a record.

This returns None if the schema is not a record.

The default implementation has to do a lot of extra work, so it is strongly recommended to implement this function when manually implementing this trait.

fn field_default() -> Option<Value>

The default value of this type when used for a record field.

None means no default value, which is also the default implementation.

Implementations of this trait provided by this crate return None except for Option<T> which returns Some(serde_json::Value::Null).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.