apache_avro

Struct GenericSingleObjectWriter

pub struct GenericSingleObjectWriter { /* private fields */ }

Writer that encodes messages according to the single object encoding v1 spec Uses an API similar to the current File Writer Writes all object bytes at once, and drains internal buffer

Implementations

impl GenericSingleObjectWriter

pub fn new_with_capacity( schema: &Schema, initial_buffer_cap: usize, ) -> AvroResult<GenericSingleObjectWriter>

pub fn new_with_capacity_and_header_builder<HB: HeaderBuilder>( schema: &Schema, initial_buffer_cap: usize, header_builder: HB, ) -> AvroResult<GenericSingleObjectWriter>

pub fn write_value_ref<W: Write>( &mut self, v: &Value, writer: &mut W, ) -> AvroResult<usize>

Write the referenced Value to the provided Write object. Returns a result with the number of bytes written including the header

pub fn write_value<W: Write>( &mut self, v: Value, writer: &mut W, ) -> AvroResult<usize>

Write the Value to the provided Write object. Returns a result with the number of bytes written including the header