apache_avro

Struct WriterBuilder

pub struct WriterBuilder<'a, W: Write, S: State = Empty> { /* private fields */ }

Use builder syntax to set the inputs and finish with build().

Implementations

impl<'a, W: Write, S: State> WriterBuilder<'a, W, S>

pub fn build(self) -> AvroResult<Writer<'a, W>>
where S: IsComplete,

Finishes building and performs the requested action.

pub fn schema(self, value: &'a Schema) -> WriterBuilder<'a, W, SetSchema<S>>
where S::Schema: IsUnset,

Required.

pub fn schemata( self, value: Vec<&'a Schema>, ) -> WriterBuilder<'a, W, SetSchemata<S>>
where S::Schemata: IsUnset,

Optional (Some / Option setters).

pub fn maybe_schemata( self, value: Option<Vec<&'a Schema>>, ) -> WriterBuilder<'a, W, SetSchemata<S>>
where S::Schemata: IsUnset,

Optional (Some / Option setters).

pub fn writer(self, value: W) -> WriterBuilder<'a, W, SetWriter<S>>
where S::Writer: IsUnset,

Required.

pub fn codec(self, value: Codec) -> WriterBuilder<'a, W, SetCodec<S>>
where S::Codec: IsUnset,

Optional (Some / Option setters). Default: Codec::Null.

pub fn maybe_codec( self, value: Option<Codec>, ) -> WriterBuilder<'a, W, SetCodec<S>>
where S::Codec: IsUnset,

Optional (Some / Option setters). Default: Codec::Null.

pub fn block_size(self, value: usize) -> WriterBuilder<'a, W, SetBlockSize<S>>
where S::BlockSize: IsUnset,

Optional (Some / Option setters). Default: DEFAULT_BLOCK_SIZE.

pub fn maybe_block_size( self, value: Option<usize>, ) -> WriterBuilder<'a, W, SetBlockSize<S>>
where S::BlockSize: IsUnset,

Optional (Some / Option setters). Default: DEFAULT_BLOCK_SIZE.

pub fn marker(self, value: [u8; 16]) -> WriterBuilder<'a, W, SetMarker<S>>
where S::Marker: IsUnset,

Optional (Some / Option setters). Default: generate_sync_marker().

pub fn maybe_marker( self, value: Option<[u8; 16]>, ) -> WriterBuilder<'a, W, SetMarker<S>>
where S::Marker: IsUnset,

Optional (Some / Option setters). Default: generate_sync_marker().

pub fn has_header(self, value: bool) -> WriterBuilder<'a, W, SetHasHeader<S>>
where S::HasHeader: IsUnset,

Optional (Some / Option setters). Default: false.

Has the header already been written.

To disable writing the header, this can be set to true.

pub fn maybe_has_header( self, value: Option<bool>, ) -> WriterBuilder<'a, W, SetHasHeader<S>>
where S::HasHeader: IsUnset,

Optional (Some / Option setters). Default: false.

Has the header already been written.

To disable writing the header, this can be set to true.

pub fn user_metadata( self, value: HashMap<String, Value>, ) -> WriterBuilder<'a, W, SetUserMetadata<S>>
where S::UserMetadata: IsUnset,

Optional (Some / Option setters). Default: <HashMap<String, Value> as Default>::default().

pub fn maybe_user_metadata( self, value: Option<HashMap<String, Value>>, ) -> WriterBuilder<'a, W, SetUserMetadata<S>>
where S::UserMetadata: IsUnset,

Optional (Some / Option setters). Default: <HashMap<String, Value> as Default>::default().