apache_avro::types

Struct Record

pub struct Record<'a> {
    pub fields: Vec<(String, Value)>,
    /* private fields */
}

Utility interface to build Value::Record objects.

Fields
fields: Vec<(String, Value)>

List of fields contained in the record. Ordered according to the fields in the schema given to create this Record object. Any unset field defaults to Value::Null.

Implementations

impl Record<'_>

pub fn new(schema: &Schema) -> Option<Record<'_>>

Create a Record given a Schema.

If the Schema is not a Schema::Record variant, None will be returned.

pub fn put<V>(&mut self, field: &str, value: V)
where V: Into<Value>,

Put a compatible value (implementing the ToAvro trait) in the Record for a given field name.

NOTE Only ensure that the field name is present in the Schema given when creating this Record. Does not perform any schema validation.

pub fn get(&self, field: &str) -> Option<&Value>

Get the value for a given field name. Returns None if the field is not present in the schema