apache_avro

Module util

Utility functions, like configuring various global settings.

Constants

DEFAULT_MAX_ALLOCATION_BYTES

Maximum number of bytes that can be allocated when decoding Avro-encoded values.

DEFAULT_SERDE_HUMAN_READABLE

Whether to set serialization & deserialization traits as human_readable or not.

Functions

max_allocation_bytes

Set the maximum number of bytes that can be allocated when decoding data.

set_serde_human_readable

Set whether the serializer and deserializer should indicate to types that the format is human-readable.