apache_avro::util

Function max_allocation_bytes

pub fn max_allocation_bytes(num_bytes: usize) -> usize

Set the maximum number of bytes that can be allocated when decoding data.

This function only changes the setting once. On subsequent calls the value will stay the same as the first time it is called. It is automatically called on first allocation and defaults to DEFAULT_MAX_ALLOCATION_BYTES.

Returns

The configured maximum, which might be different from what the function was called with if the value was already set before.