apache_avro::writer

Module datum

Structs

GenericDatumWriter

Writer for writing raw Avro data.

GenericDatumWriterBuilder

Use builder syntax to set the inputs and finish with build().

Functions

to_avro_datumDeprecated

Deprecated. Use GenericDatumWriter instead.

to_avro_datum_schemataDeprecated

Deprecated. Use GenericDatumWriter instead.

write_avro_datum_ref

Write the referenced Serializeable object to the provided Write object.