apache_avro::writer::single_object

Struct SpecificSingleObjectWriterBuilder

pub struct SpecificSingleObjectWriterBuilder<T, S: State = Empty>
where
    T: AvroSchema,
{ /* private fields */ }

Use builder syntax to set the inputs and finish with build().

Implementations

impl<T, S: State> SpecificSingleObjectWriterBuilder<T, S>
where T: AvroSchema,

pub fn build(self) -> SpecificSingleObjectWriter<T>
where S: IsComplete,

Finish building and return the requested object

pub fn resolved( self, schema: Schema, ) -> Result<SpecificSingleObjectWriterBuilder<T, SetResolved<S>>, Error>
where S::Resolved: IsUnset,

Optional (Some / Option setters). Default:

ResolvedOwnedSchema::new(T::get_schema())
    .expect("AvroSchema implementation should create valid schemas")

pub fn maybe_resolved( self, value: Option<Schema>, ) -> Result<SpecificSingleObjectWriterBuilder<T, SetResolved<S>>, Error>
where S::Resolved: IsUnset,

Optional (Some / Option setters). Default:

ResolvedOwnedSchema::new(T::get_schema())
    .expect("AvroSchema implementation should create valid schemas")

pub fn header( self, header_builder: impl HeaderBuilder, ) -> SpecificSingleObjectWriterBuilder<T, SetHeader<S>>
where S::Header: IsUnset,

Optional (Some / Option setters). Default:

RabinFingerprintHeader::from_schema(resolved.get_root_schema())
    .build_header()

pub fn maybe_header( self, value: Option<impl HeaderBuilder>, ) -> SpecificSingleObjectWriterBuilder<T, SetHeader<S>>
where S::Header: IsUnset,

Optional (Some / Option setters). Default:

RabinFingerprintHeader::from_schema(resolved.get_root_schema())
    .build_header()

pub fn human_readable( self, value: bool, ) -> SpecificSingleObjectWriterBuilder<T, SetHumanReadable<S>>
where S::HumanReadable: IsUnset,

Optional (Some / Option setters). Default: is_human_readable().

Should Serialize implementations pick a human readable representation.

It is recommended to set this to false.

pub fn maybe_human_readable( self, value: Option<bool>, ) -> SpecificSingleObjectWriterBuilder<T, SetHumanReadable<S>>
where S::HumanReadable: IsUnset,

Optional (Some / Option setters). Default: is_human_readable().

Should Serialize implementations pick a human readable representation.

It is recommended to set this to false.

pub fn target_block_size( self, value: usize, ) -> SpecificSingleObjectWriterBuilder<T, SetTargetBlockSize<S>>
where S::TargetBlockSize: IsUnset,

Optional (Some / Option setters). At what block size to start a new block (for arrays and maps).

This is a minimum value, the block size will always be larger than this except for the last block.

When set to None all values will be written in a single block. This can be faster as no intermediate buffer is used, but seeking through written data will be slower.

pub fn maybe_target_block_size( self, value: Option<usize>, ) -> SpecificSingleObjectWriterBuilder<T, SetTargetBlockSize<S>>
where S::TargetBlockSize: IsUnset,

Optional (Some / Option setters). At what block size to start a new block (for arrays and maps).

This is a minimum value, the block size will always be larger than this except for the last block.

When set to None all values will be written in a single block. This can be faster as no intermediate buffer is used, but seeking through written data will be slower.