apache_avro

Module schema

Logic for parsing and interacting with schemas in Avro format.

Structs

Alias

Newtype pattern for Name to better control the serde_json::Value representation. Aliases are serialized as an array of plain strings in the JSON representation.

ArraySchema

DecimalSchema

A description of a Decimal schema.

EnumSchema

A description of an Enum schema.

EnumSchemaBuilder

Use builder syntax to set the inputs and finish with build().

FixedSchema

A description of a Fixed schema.

FixedSchemaBuilder

Use builder syntax to set the inputs and finish with build().

MapSchema

Name

Represents names for record, enum and fixed Avro schemas.

RecordField

Represents a field in a record Avro schema.

RecordFieldBuilder

Use builder syntax to set the inputs and finish with build().

RecordSchema

A description of a Record schema.

RecordSchemaBuilder

Use builder syntax to set the inputs and finish with build().

ResolvedSchema

SchemaFingerprint

Represents an Avro schema fingerprint.

UnionSchema

A description of a Union schema

Enums

InnerDecimalSchema

The inner schema of the Decimal type.

RecordFieldOrder

Represents any valid order for a field in a record Avro schema.

Schema

Represents any valid Avro schema More information about Avro schemas can be found in the Avro Specification

SchemaKind

UuidSchema

The inner schema of the Uuid type.

Type Aliases

Aliases

Represents the aliases for Named Schema

Documentation

Represents documentation for complex Avro schemas.

Names

Represents Schema lookup within a schema env

NamesRef

Represents Schema lookup within a schema

Namespace

Represents the namespace for Named Schema