Enum apache_avro::schema::SchemaKind

pub enum SchemaKind {

    Null,
    Boolean,
    Int,
    Long,
    Float,
    Double,
    Bytes,
    String,
    Array,
    Map,
    Union,
    Record,
    Enum,
    Fixed,
    Decimal,
    Uuid,
    Date,
    TimeMillis,
    TimeMicros,
    TimestampMillis,
    TimestampMicros,
    Duration,
    Ref,

}

Auto-generated discriminant enum variants

Variants

Null

A null Avro schema.

Boolean

A boolean Avro schema.

Int

An int Avro schema.

Long

A long Avro schema.

Float

A float Avro schema.

Double

A double Avro schema.

Bytes

A bytes Avro schema. Bytes represents a sequence of 8-bit unsigned bytes.

String

A string Avro schema. String represents a unicode character sequence.

Array

A array Avro schema. Avro arrays are required to have the same type for each element. This variant holds the Schema for the array element type.

Map

A map Avro schema. Map holds a pointer to the Schema of its values, which must all be the same schema. Map keys are assumed to be string.

Union

A union Avro schema.

Record

A record Avro schema.

The lookup table maps field names to their position in the Vec of fields.

Enum

An enum Avro schema.

Fixed

A fixed Avro schema.

Decimal

Logical type which represents Decimal values. The underlying type is serialized and deserialized as Schema::Bytes or Schema::Fixed.

scale defaults to 0 and is an integer greater than or equal to 0 and precision is an integer greater than 0.

Uuid

A universally unique identifier, annotating a string.

Date

Logical type which represents the number of days since the unix epoch. Serialization format is Schema::Int.

TimeMillis

The time of day in number of milliseconds after midnight with no reference any calendar, time zone or date in particular.

TimeMicros

The time of day in number of microseconds after midnight with no reference any calendar, time zone or date in particular.

TimestampMillis

An instant in time represented as the number of milliseconds after the UNIX epoch.

TimestampMicros

An instant in time represented as the number of microseconds after the UNIX epoch.

Duration

An amount of time defined by a number of months, days and milliseconds.

Ref

Implementations

impl SchemaKind

pub fn is_primitive(self) -> bool

pub fn is_named(self) -> bool