apache_avro::serde

Function to_value

pub fn to_value<S: Serialize>(value: S) -> Result<Value, Error>

Interpret a serializeable instance as a Value.

This conversion can fail if the value is not valid as per the Avro specification. e.g: HashMap with non-string keys.

This function does not work if S has any fields (recursively) that have the #[serde(flatten)] attribute. Please use Writer::append_ser if that’s the case.