apache_avro::validator

Function set_enum_symbol_name_validator

pub fn set_enum_symbol_name_validator(
    validator: Box<dyn EnumSymbolNameValidator + Send + Sync>,
) -> Result<(), Box<dyn EnumSymbolNameValidator + Send + Sync>>

Sets a custom enum symbol name validator.

Returns a unit if the registration was successful or the already registered validator if the registration failed.

Note: This function must be called before parsing any schema because this will register the default validator and the registration is one time only!